pre refactor
