+++ backend/auth_service/app/main.py
@@ -1,10 +1,6 @@
 from fastapi import FastAPI
 from api.v1.api.router import router as auth
-from fastapi import middleware
-
 
 app = FastAPI() 
 
 app.include_router(auth)
-
-app.middleware()
