+++ backend/auth_service/app/models/auth.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, EmailStr
+
+class UserLoginShema(BaseModel):
+    email: EmailStr
+    password: str
