+++ backend/auth_service/app/repository/user_repository.py
@@ -1,12 +1,8 @@
-from app.database import async_session_maker
+from app.repository.database import async_session_maker
 from pydantic import EmailStr
 from sqlalchemy import text
-from app.config import Settings
+from app.core.config import Settings
 from app.core.httpexceptions import UserNotFoundException
-from schema.user import UserLoginShema, UserSchemaForDump
-import jwt
-from datetime import datetime, timedelta, timezone
-from fastapi import HTTPException
 
 
 class UserRepository:
